requirements.txt
torch>=1.13.1
diffusers>=0.25.0
transformers
accelerate
xformers
safetensors
numpy
pandas
Pillow
PyYAML
matplotlib
